requirements.txt
ansicolors
pyinstaller